% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_metadata}
\alias{all_metadata}
\title{Meta Data for Acute Lymphoblastic Leukaemia (ALL) Microarry Data Set}
\format{
A data frame with 75 samples (rows) and 3 variables (columns):
\describe{
\item{Patient_ID}{ID of the Patient}
\item{Class}{Whether the sample is a "BCR" or "E2A" sample; "BCR" stands for the BCR-ABL leukemia subtype
while "E2A" stands for the E2A-PBX1 leukemia subtype}
\item{Batch}{Batch of the Sample (R/Y; R stands for Ross et al., Y stands for Yeoh et al.)}
}
}
\usage{
data(all_metadata)
}
\description{
Refer to \code{\link{all}} for details on the Acute Lymphoblastic Leukaemia (ALL) microarry data set.
}
\details{
Contains the "Patient ID","Class" and "Batch" of each sample.
}
\references{
Ross ME, Mahfouz R, Onciu M, Liu H-C, Zhou X, Song G, et al. Gene expression profiling of pediatric acute myelogenous leukemia Blood. 2004; 104:3679-87.

Yeoh E-J, Ross ME, Shurtleff SA, Williams WK, Patel D, Mahfouz R, et al. Classification, subtype discovery, and prediction of outcome in pediatric acute lymphoblastic leukemia by gene expression profiling Cancer Cell. 2002; 1:133-43.
}
\keyword{datasets}
