% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmd}
\alias{dmd}
\title{Duchenne Muscular Dystrophy (DMD) Microarray Data Set}
\format{
A data frame with 60 samples (columns) and 8813 ENSEMBL ID variables (rows):
}
\usage{
data(dmd)
}
\description{
Unprocessed data set (no batch correction performed) formed from the combination
of 2 independently-derived DMD data sets, Haslett et al. and Pescatori et al.
}
\details{
Both data sets were combined into a single data set with the following procedure:
\enumerate{
\item All probes of both data sets were converted to ENSEMBL IDs using biomaRt.
\item To ensure a one-to-one mapping between the probes and ENSEMBL IDs in both data
sets, all probes with no ENSEMBL ID were removed. Probes with multiple ENSEMBL
IDs were replaced by the ENSEMBL ID with the smallest value (ENSEMBL IDs were
ordered using the default order function and all ENSEMBL IDs after the first
ENSEMBL ID was removed). We took the median values of probes sharing the same
ENSEMBL ID. After this procedure, both data sets would consist of unique ENSEMBL
ID variables.
\item To join both data sets without any null values or data imputation (since both
data sets may not have the same number and type of ENSEMBL IDs), we took the
intersection of ENSEMBL IDs between both data sets. This set of ENSEMBL IDs
would be the ENSEMBL IDs of the joined data set.
\item Both data sets were joined along the shared set of ENSEMBL IDs.
}
}
\references{
Haslett JN, Sanoudou D, Kho AT, Bennett RR, Greenberg SA, Kohane IS, et al. Gene expression comparison of biopsies from Duchenne muscular dystrophy (DMD) and normal skeletal muscle Proceedings of the National Academy of Sciences. 2002; 99:15000-5.

Pescatori M, Broccolini A, Minetti C, Bertini E, Bruno C, D'amico A, et al. Gene expression profiling in the early phases of DMD: a constant molecular signature characterizes DMD muscle from early postnatal life throughout disease progression The FASEB Journal. 2007; 21:1210-26.
}
\keyword{datasets}
