% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmd_metadata}
\alias{dmd_metadata}
\title{Meta Data for Duchenne Muscular Dystrophy (DMD) Microarray Data Set}
\format{
A data frame with 60 samples (rows) and 3 variables (columns):
\describe{
\item{Patient_ID}{ID of the Patient}
\item{Class}{Whether the sample is a "DMD" or "NOR" sample}
\item{Batch}{Batch of the Sample (H/P; H stands for Haslett et al., P stands for Pescatori et al.)}
}
}
\usage{
data(dmd_metadata)
}
\description{
Refer to \code{\link{dmd}} for details on the DMD data set.
}
\details{
Contains the "Patient ID","Class" and "Batch" of each sample.
}
\references{
Haslett JN, Sanoudou D, Kho AT, Bennett RR, Greenberg SA, Kohane IS, et al. Gene expression comparison of biopsies from Duchenne muscular dystrophy (DMD) and normal skeletal muscle Proceedings of the National Academy of Sciences. 2002; 99:15000-5.

Pescatori M, Broccolini A, Minetti C, Bertini E, Bruno C, D'amico A, et al. Gene expression profiling in the early phases of DMD: a constant molecular signature characterizes DMD muscle from early postnatal life throughout disease progression The FASEB Journal. 2007; 21:1210-26.
}
\keyword{datasets}
