% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doppelgangerVerification.R
\name{doppelgangerVerification}
\alias{doppelgangerVerification}
\title{Trains 12*numberOfTrainingAndValidationSets KNN classifiers to verify the
functionality of PPCC}
\usage{
doppelgangerVerification(
  experimentPlanFilename,
  raw_data,
  meta_data,
  featureSetPortion = 0.1,
  seednum = 2021,
  separator = "\\\\."
)
}
\arguments{
\item{experimentPlanFilename}{Name of file containing csv experiment plan}

\item{raw_data}{Data frame where each column is a sample and each row is a variable}

\item{featureSetPortion}{Proportion of variables to be used for feature set generation}

\item{seednum}{Seed number for random feature set generation}

\item{separator}{The character separating the name of the training_validation pair
e.g. "0 Doppel" from the "train", "valid" label. Name of each column should be in
format "0 Doppel.train" if . is used as separator.}

\item{metadata}{File containing metadata}
}
\value{
Validation Accuracies
}
\description{
This function verifies the functionality of PPCC
}
\examples{
doppelgangerVerification()
}
