% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPPCCDoppelgangers.R
\name{getPPCCDoppelgangers}
\alias{getPPCCDoppelgangers}
\title{A function to get PPCC correlation matrix and data frame and list of doppelgangers}
\usage{
getPPCCDoppelgangers(raw_data, meta_data)
}
\arguments{
\item{raw_data}{Data frame where each column is a sample and each row is a variable}

\item{meta_data}{Data frame with the columns "Class", "Patient_ID", "Batch"
indicating the class, patient id and batch of the sample respectively and each row
is a sample}
}
\value{
A list containing the PPCC matrix and data frame and a list of
doppelgangers identified
}
\description{
This function gets PPCC between a samples of 2 different data sets/batches and
identifies doppelgangers between them
}
\examples{
getPPCCDoppelgangers()
}
