% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelClassPatient.R
\name{labelClassPatient}
\alias{labelClassPatient}
\title{A function to label rows of PPCC data frame according to their class and patient id}
\usage{
labelClassPatient(ppcc_df, meta_data)
}
\arguments{
\item{ppcc_df}{Data frame with "Sample1", "Sample2", "PPCC" as column names}

\item{meta_data}{Data frame with the columns "Class" and "Patient_ID"
indicating the class and patient id of the sample respectively and each row
is a sample}
}
\value{
Labeled PPCC data frame
labelClassPatient()
}
\description{
This function labels PPCC data frame rows according to their class and patient
similarity
}
