% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelClassPatientPerPair.R
\name{labelClassPatientPerPair}
\alias{labelClassPatientPerPair}
\title{A function to label a pair of samples to their class and patient id}
\usage{
labelClassPatientPerPair(
  sample1Class,
  sample1Patient,
  sample2Class,
  sample2Patient
)
}
\arguments{
\item{sample1Class}{Class of sample 1}

\item{sample1Patient}{Patient_ID of sample 1}

\item{sample2Class}{Class of sample 2}

\item{sample2Patient}{Patient_ID of sample 2}
}
\value{
Class Patient Label of the Sample Pair
labelClassPatientPerPair()
}
\description{
This function labels a pair of samples to their class and patient id
}
