% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leuk_metadata}
\alias{leuk_metadata}
\title{Meta Data for Leukemia Microarry Data Set}
\format{
A data frame with 60 samples (rows) and 3 variables (columns):
\describe{
\item{Patient_ID}{ID of the Patient}
\item{Class}{Whether the sample is a "ALL" or "AML" sample; "ALL" stands for Acute Lymphoblastic
Leukaemia while "AML" stands for Acute Myeloid Leukaemia}
\item{Batch}{Batch of the Sample (A/G; A stands for Armstrong et al., G stands for Golub et al.)}
}
}
\usage{
data(leuk_metadata)
}
\description{
Refer to \code{\link{leuk}} for details on the Leukemia Microarry data set.
}
\details{
Contains the "Patient ID","Class" and "Batch" of each sample.
}
\references{
Golub TR, Slonim DK, Tamayo P, Huard C, Gaasenbeek M, Mesirov JP, et al. Molecular classification of cancer: class discovery and class prediction by gene expression monitoring Science. 1999; 286:531-7.

Armstrong SA, Staunton JE, Silverman LB, Pieters R, den Boer ML, Minden MD, et al. MLL translocations specify a distinct gene expression profile that distinguishes a unique leukemia Nat Genet. 2002; 30:41-7.
}
\keyword{datasets}
