% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadExperimentPlan.R
\name{loadExperimentPlan}
\alias{loadExperimentPlan}
\title{A function to load a csv experiment plan into list form}
\usage{
loadExperimentPlan(filename, raw_data, separator = "\\\\.")
}
\arguments{
\item{filename}{Name of csv file}

\item{raw_data}{Data frame where each column is a sample and each row is a variable}

\item{separator}{The character separating the name of the training_validation pair
e.g. "0 Doppel" from the "train", "valid" label. Name of each column should be in
format "0 Doppel.train" if . is used as separator}
}
\value{
Experiment plan in list form
loadExperimentPlan()
}
\description{
This function load a csv experiment plan into list form
}
