% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rc}
\alias{rc}
\title{Renal Cell Carcinoma Proteomics Data Set}
\format{
A data frame with 36 samples (columns) and 3126 UniProt ID variables (rows):
}
\source{
\url{https://dx.doi.org/10.1038\%2Fnm.3807}
}
\usage{
data(rc)
}
\description{
A benchmark proteomics data taken from the NetProt software library.The renal cell carcinoma (RCC) study of Guo et al.
comprises 24 proteomics runs originating from 6 pairs of non-tumorous and tumorous clear-cell renal carcinoma (ccRCC)
tissues, 4 proteomics runs originating from a pair of non-tumorous and tumorous chromophobe renal carcinoma (chRCC)
tissues and 8 proteomics run originating from two pairs of non-tumorous and tumorous papillary renal carcinoma (ppRCC)
tissues, in duplicates.
}
\references{
Guo T, Kouvonen P, Koh CC et al. Rapid mass spectrometric conversion of tissue biopsy samples into permanent quantitative digital proteome maps, Nature medicine 2015;21:407-413.
}
\keyword{datasets}
