% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseVerificationResults.R
\name{visualiseVerificationResults}
\alias{visualiseVerificationResults}
\title{Displays accuracies from functionality analysis}
\usage{
visualiseVerificationResults(
  functionalityResults,
  originalTrainValidNames = c(),
  newTrainValidNames = c()
)
}
\arguments{
\item{ppccDoppelgangerResults}{List returned from  \code{\link{doppelgangerFunctionalityVerification}}}

\item{originalTrainVaidNames}{Original names of training and validation pairs}

\item{newTrainVaidNames}{New names of training and validation pairs}
}
\description{
Displays results from \code{\link{doppelgangerFunctionalityVerification}}
}
\examples{
visualiseVerificationResults()
}
