% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rc_metadata}
\alias{rc_metadata}
\title{Meta Data for Renal Cell Carcinoma Proteomics Data Set}
\format{
A data frame with 36 samples (rows) and 5 variables (columns):
\describe{
\item{Class}{Whether the sample is a "Tumor" or "Normal" tissue sample}
\item{Tumour_Histological_Type}{Histological Type of the Tumour; (cc: Clear cell, p: Papillary, ch: Chromophobe)}
\item{Patient_ID}{ID of the Patient (1-8)}
\item{Sample_ID}{ID of the Sample (1-36)}
\item{Batch}{Batch number of the Sample (1-2)}
}
}
\source{
\url{https://dx.doi.org/10.1038\%2Fnm.3807}
}
\usage{
data(rc)
}
\description{
Contains the "Class"
}
\references{
Guo T, Kouvonen P, Koh CC et al. Rapid mass spectrometric conversion of tissue biopsy samples into permanent quantitative digital proteome maps, Nature medicine 2015;21:407-413.
}
\keyword{datasets}
