% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyDoppelgangers.R
\name{verifyDoppelgangers}
\alias{verifyDoppelgangers}
\title{Verifies the functionality of Doppelgangers}
\usage{
verifyDoppelgangers(
  experimentPlanFilename,
  raw_data,
  meta_data,
  featureSetPortion = 0.1,
  seednum = 2021,
  separator = "\\\\.",
  do.batch.corr = TRUE,
  k = 5
)
}
\arguments{
\item{experimentPlanFilename}{Name of file containing csv experiment plan.
The csv file has a header with the names of the training_validation sets (e.g. "Doppel_0.train" or "Doppel_0.valid").
In each column (e.g. "Doppel_0.train" column), we include the names of all samples included in this training/validation set.}

\item{featureSetPortion}{Proportion of variables to be used for feature set generation}

\item{seednum}{Seed number for random feature set generation}

\item{separator}{The character separating the name of the training_validation pair
e.g. "0 Doppel" from the "train", "valid" label. Name of each column should be in
format "0 Doppel.train" if . is used as separator}

\item{ppccDoppelgangerResults}{List returned from \code{\link{getPPCCDoppelgangers}}}

\item{metadata}{File containing metadata}
}
\value{
Validation Accuracies
}
\description{
The user constructs a csv file with with training-validation set pairs
ideally incrementing the number of Doppelgangers between training and validation sets.
For each training-validation set pair, 12 models with different feature sets will be trained.
10 random feature sets and 2 features sets of highest and lowest variance would be generated.
If an increase in validation accuracy of the 10 random models with increasing number of doppelgangers
can be observed, we can conclude that the doppelgangers included are functional doppelgangers.
}
\details{
\strong{Troubleshooting tips:}
\itemize{
\item Ensure all the headers have no spaces.
\item If excel is used for planning, save the spreadsheet as "CSV (MS-DOS) (*.csv)"
\item Use the exact label "train" and "valid" (take note of capital letters)
\item Ensure the separator does not exist in the name of the training-validation set (E.g. Doppel.0 is not allowed)
\item Try to put both training-validation columns beside each other and leave no column gaps
\item Refer to the csv file in the tutorial on the GitHub README.
}
}
\examples{
verificationResults = verifyDoppelgangers(
experimentPlanFilename = "tutorial/experimentPlan.csv",
raw_data = rc,
meta_data = rc_metadata)
}
