% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisePPCCDoppelgangers.R
\name{visualisePPCCDoppelgangers}
\alias{visualisePPCCDoppelgangers}
\title{A function to visualize PPCC data doppelgangers on a scatter plot}
\usage{
visualisePPCCDoppelgangers(ppccDoppelgangerResults)
}
\arguments{
\item{ppccDoppelgangerResults}{List returned from \code{\link{getPPCCDoppelgangers}}}
}
\description{
This function gets visualizes PPCC data doppelgangers on a scatter plot
}
\examples{
visualisePPCCDoppelgangers()
}
