% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisePPCCDoppelgangers.R
\name{visualisePPCCDoppelgangers}
\alias{visualisePPCCDoppelgangers}
\title{A function to visualize PPCC data doppelgangers on a scatter plot}
\usage{
visualisePPCCDoppelgangers(ppcc_doppelganger_results)
}
\arguments{
\item{ppcc_doppelganger_results}{List returned from \code{\link{getPPCCDoppelgangers}}}
}
\description{
This function visualizes PPCC data doppelgangers from \code{\link{getPPCCDoppelganger}}
on adjacent univariate scatter plot.
}
\examples{
visualisePPCCDoppelgangers(ppcc_doppelganger_results = ppcc_doppelganger_results)
}
