% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseVerificationResults.R
\name{visualiseVerificationResults}
\alias{visualiseVerificationResults}
\title{Displays accuracies from functionality analysis}
\usage{
visualiseVerificationResults(
  verification_results,
  original_train_valid_names = c(),
  new_train_valid_names = c()
)
}
\arguments{
\item{verification_results}{List returned from  \code{\link{verifyDoppelgangers}}}

\item{original_train_valid_names}{Original names of training and validation pairs}

\item{new_train_valid_names}{New names of training and validation pairs}
}
\description{
Displays results from \code{\link{verifyDoppelgangers}}
}
\examples{
ori_train_valid_names = c("Doppel_0","Doppel_2", "Doppel_4", "Doppel_6", "Doppel_8", "Neg_Con", "Pos_Con")

new_train_valid_names = c("0 Doppel", "2 Doppel", "4 Doppel", "6 Doppel", "8 Doppel", "Binomial", "Perfect Leakage")

visualiseVerificationResults( verification_results,
                              ori_train_valid_names,
                              new_train_valid_names)
}
